clean up the scan window overlay
