+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Bugs fixed:
 
 Improvements:
 * Added a basic barcode overlay widget, for use with the camera preview.
+* Added a basic scan window overlay widget, for use with the camera preview.
 * Update the bundled MLKit model for Android to version `17.3.0`.
 * Added documentation in places where it was missing.
 * Added `color` and `style` properties to the `BarcodePainter` widget.
